+++ src/main.cpp
@@ -8,40 +8,9 @@
 #include "behaviortree_cpp/loggers/groot2_publisher.h"
 #include "ros/ros.h"
 
-#include "sonia_bt_runner/conditions/SoniaConditions.hpp"
-#include "sonia_bt_runner/sync_actions/SoniaSyncActions.hpp"
-#include "sonia_bt_runner/stateful_actions/SoniaStatefulActions.hpp"
+#include "sonia_bt_runner/SoniaNodes.hpp"
 
 using namespace BT;
-
-static void registerConditionNodes(BT::BehaviorTreeFactory &factory)
-{
-    factory.registerNodeType<MissionSwitchStatus>("MissionSwitchStatus");
-}
-static void registerSyncActionNodes(BT::BehaviorTreeFactory &factory)
-{
-    factory.registerNodeType<DropperActionNode>("DropperActionNode");
-    factory.registerNodeType<InitTraj>("InitTraj");
-    factory.registerNodeType<AddPoseToTraj>("AddPoseToTraj");
-}
-
-static void registerStatefulActionNodes(BT::BehaviorTreeFactory &factory)
-{
-    factory.registerNodeType<SendTrajToPlanner>("SendTrajToPlanner");
-    factory.registerNodeType<WaitTargetReached>("WaitTargetReached");
-}
-
-static void registerNodes(BT::BehaviorTreeFactory &factory)
-{
-    registerConditionNodes(factory);
-    registerSyncActionNodes(factory);
-    registerStatefulActionNodes(factory);
-}
-
-void Spin()
-{
-    ros::spin();
-}
 
 int main(int argc, char *argv[])
 {
@@ -53,18 +22,6 @@ int main(int argc, char *argv[])
     BT::BehaviorTreeFactory factory;
     registerNodes(factory);
     
-    if ((std::string)argv[1] == "gen_models"){
-        if (argc != 3) {
-            return EXIT_FAILURE;
-        }
-        std::string xml_models = BT::writeTreeNodesModelXML(factory);
-        printf("%s\n", xml_models.c_str());
-        std::ofstream myfile;
-        myfile.open(argv[2]);
-        myfile << xml_models.c_str();
-        myfile.close();
-        return 0;
-    }
     // factory.registerNodeType<DropperActionNode>("DropperActionNode");
 
     std::string name = argv[1];
